+++ src/state.rs
@@ -3,8 +3,7 @@ use std::{self,
           sync::{Arc, Mutex}};
 
 use super::evm::{get_contract_address, GAS_LIMIT, SPEC};
-use ekiden_core::{self, error::Result, mrae::sivaessha2::NONCE_SIZE};
-use ekiden_keymanager_common::confidential;
+use ekiden_core::{self, error::Result};
 use ekiden_storage_base::StorageBackend;
 use ekiden_trusted::db::{Database, DatabaseHandle};
 use ethcore::{self,
@@ -15,7 +14,7 @@ use ethcore::{self,
               filter::Filter as EthcoreFilter,
               header::Header,
               kvdb::{self, KeyValueDB},
-              state::{backend::Wrapped as WrappedBackend, Encrypter, KeyManager},
+              state::backend::Wrapped as WrappedBackend,
               transaction::Action,
               types::{ids::BlockId,
                       log_entry::{LocalizedLogEntry, LogEntry},
@@ -23,7 +22,12 @@ use ethcore::{self,
                       BlockNumber}};
 use ethereum_api::{BlockId as EkidenBlockId, Filter, Log, Receipt, Transaction};
 use ethereum_types::{Address, H256, U256};
-use runtime_ethereum_common::{get_factories, Backend, BlockchainStateDb, State, StorageHashDB};
+use runtime_ethereum_common::{confidential::{Encrypter, KeyManager},
+                              get_factories,
+                              Backend,
+                              BlockchainStateDb,
+                              State,
+                              StorageHashDB};
 
 lazy_static! {
     static ref GLOBAL_CACHE: Mutex<Option<Cache>> = Mutex::new(None);
@@ -116,8 +120,8 @@ impl Cache {
             root,
             U256::zero(), /* account_start_nonce */
             get_factories(),
-            Some(Box::new(_Encrypter)),
-            Some(Box::new(_KeyManager)),
+            Some(Box::new(Encrypter)),
+            Some(Box::new(KeyManager)),
         )?)
     }
 
@@ -135,8 +139,8 @@ impl Cache {
             vec![],                           /* extra data */
             true,                             /* is epoch_begin */
             &mut Vec::new().into_iter(),      /* ancestry */
-            Some(Box::new(_Encrypter)),
-            Some(Box::new(_KeyManager)),
+            Some(Box::new(Encrypter)),
+            Some(Box::new(KeyManager)),
         )?)
     }
 
@@ -329,47 +333,6 @@ impl Cache {
 
     pub fn best_block_header(&self) -> Header {
         self.chain.best_block_header()
-    }
-}
-
-/// Implementation of the parity KeyManager trait to inject into the ConfidentialVm.
-struct _KeyManager;
-impl KeyManager for _KeyManager {
-    fn long_term_public_key(&self, contract: Address) -> Vec<u8> {
-        let (pk, _sk) = confidential::default_contract_keys();
-        pk.to_vec()
-    }
-}
-
-/// Implementation of the Encrypter trait to inject into parity for confidential contracts.
-struct _Encrypter;
-impl Encrypter for _Encrypter {
-    fn encrypt(
-        &self,
-        plaintext: Vec<u8>,
-        peer_public_key: Vec<u8>,
-    ) -> std::result::Result<Vec<u8>, String> {
-        if peer_public_key.len() < 32 {
-            return Err("public keys must be 32 bytes long".to_string());
-        }
-        // just generate arbitrary nonce for now (change this to a random nonce once we encrypt
-        // with an actual key manager)
-        let mut nonce = [1u8; NONCE_SIZE];
-        let mut pub_key: [u8; 32] = Default::default();
-        pub_key.copy_from_slice(&peer_public_key[..32]);
-        confidential::encrypt(plaintext, nonce.to_vec(), pub_key)
-            .map_err(|_| "could not decrypt".to_string())
-    }
-
-    /// Returns a tuple containing the nonce, public key, and plaintext
-    /// used to generate the given cypher.
-    fn decrypt(&self, cypher: Vec<u8>) -> std::result::Result<(Vec<u8>, Vec<u8>, Vec<u8>), String> {
-        let decryption = confidential::decrypt(Some(cypher)).map_err(|_| "Error".to_string())?;
-        Ok((
-            decryption.nonce,
-            decryption.peer_public_key.to_vec(),
-            decryption.plaintext,
-        ))
     }
 }
